producer/requirements.txt
kafka-python==2.0.2
sseclient==0.0.27